requirements.txt
fastapi
uvicorn
requests
python-multipart
google-api-python-client
google-auth
openai